Rename: 프로필 수정 페이지 폴더 이동
프로필 수정 페이지 폴더 이동 profile -> profile/edit
